Added grammer to remove highest and lowest dice from dice pools
Added a grammer to remove the highest (rXh, where X is the number of dice to
remove; defaults to 1 of not present) and the lowest (rXl, where X is the number of
dice to remove; defaults to 1 if not present). Full dice pool notation would looke
like `10d10r2h` to roll 10 10-sided dice and remove the two highest values.

---
type: feature