Added sequece strategy to dice pool
Added sequece strategy to dice pool to facitiliate more useful testing
scenarios. Instead of picking a value at random from the faces of the die it
selected the next one in the sequence in ascending order wrapping when the
of calls exceeds the number of faces.

Set the enviornment variable `DICE_ROLLER_STRATEGY` to `SEQUENCE` to use the
strategy and the environment variable `DICE_ROLLER_SEQUENCE_COUNT` to `0` to
reset the call count.

---
type: feature